NMS-10268: Move timeformat.api and impl modules to lib and wire it through custom.properties into osgi
