+++ container/karaf/src/main/filtered-resources/etc/custom.properties
@@ -725,6 +725,7 @@ org.osgi.framework.system.packages.extra=org.apache.karaf.branding,\
         org.opennms.features.status.api;version=${opennms.osgi.version},\
         org.opennms.features.status.api.node;version=${opennms.osgi.version},\
         org.opennms.features.status.api.node.strategy;version=${opennms.osgi.version},\
+        org.opennms.features.timeformat.api;version=${opennms.osgi.version},\
         org.opennms.netmgt.bsm.service;version=${opennms.osgi.version},\
         org.opennms.netmgt.bsm.service.model;version=${opennms.osgi.version},\
         org.opennms.netmgt.bsm.service.model.edge;version=${opennms.osgi.version},\
