+++ features/vaadin-dashlets/dashlet-features/src/main/resources/features.xml
@@ -8,7 +8,6 @@
 
   <feature name="dashlet-alarms" version="${pom.version}" description="OpenNMS :: Features :: Dashlets :: Alarms">
     <feature>opennms-core</feature>
-    <bundle>mvn:org.opennms.features.timeformat/org.opennms.features.timeformat.api/${project.version}</bundle>
     <bundle>mvn:org.opennms.features.vaadin-dashlets/dashlet-alarms/${pom.version}</bundle>
   </feature>
   <feature name="dashlet-bsm" version="${pom.version}" description="OpenNMS :: Features :: Dashlets :: BSM">
